Add hooks to separate folder
