+++ packages/website/src/routes/ReefRoutes/SurveyPoint/InfoCard/EditForm.tsx
@@ -10,7 +10,7 @@ import {
 import Alert from "@material-ui/lab/Alert";
 import { find } from "lodash";
 
-import { FormField } from "../../../../common/Forms/useFormField";
+import { FormField } from "../../../../hooks/useFormField";
 import TextField from "../../../../common/Forms/TextField";
 
 const NUMERIC_FIELD_STEP = 1 / 10 ** 15;
